requirements.txt
python-telegram-bot==13.15
peewee==3.17.0
peewee-migrate==1.12.2
python-dotenv==1.0.0
loguru==0.7.2
python-i18n==0.3.9
pyyaml==6.0.1